fixed an error where refresh tokens would not be stored

+++ database/migrations/2022_12_19_164534_create_refresh_tokens_table.php
@@ -16,7 +16,7 @@ return new class extends Migration {
             $table->uuid('token_id');
             $table->string('client_id');
             $table->string('token_hash');
-            $table->float('expires');
+            $table->bigInteger('expires');
             $table->timestamps();
         });
     }
